Restructured EOF postprocessing implementation to facilitate selection of alternative approaches.
Untested so far!

Prepared comments to show places, where eof rotation method has to be added next
