+++ tests/test_omi_calculator.py
@@ -39,34 +39,27 @@ originalOMIDataDirname = Path(os.path.abspath('')) / "testdata" / "OriginalOMI"
 eof1Dirname = originalOMIDataDirname / "eof1"
 eof2Dirname = originalOMIDataDirname / "eof2"
 origOMIPCsFilename = originalOMIDataDirname / "omi.1x.txt"
-mjoindices_reference_eofs_filename_strict = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_strict.npz"
-mjoindices_reference_pcs_filename_strict = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs_strict.txt"
-mjoindices_reference_eofs_filename_coarsegrid = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_coarsegrid.npz"
-mjoindices_reference_pcs_filename_coarsegrid = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs_coarsegrid.txt"
-mjoindices_reference_eofs_filename_strict_eofs = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_eofs_package.npz"
-mjoindices_reference_pcs_filename_strict_eofs = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs_eofs_package.txt"
+mjoindices_reference_eofs_filename_strict = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_strict.npz"
+mjoindices_reference_pcs_filename_strict = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs_strict.txt"
+mjoindices_reference_eofs_filename_coarsegrid = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_coarsegrid.npz"
+mjoindices_reference_pcs_filename_coarsegrid = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs_coarsegrid.txt"
+mjoindices_reference_eofs_filename_strict_eofs = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_eofs_package.npz"
+mjoindices_reference_pcs_filename_strict_eofs = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs_eofs_package.txt"
 mjoindices_reference_eofs_filename = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs.npz"
 mjoindices_reference_pcs_filename = Path(os.path.abspath('')) / "testdata" / "mjoindices_reference" / "PCs.txt"
+mjoindices_reference_eofs_filename_raw = Path(
+    os.path.abspath('')) / "testdata" / "mjoindices_reference" / "EOFs_raw.npz"
 original_omi_explained_variance_file = Path(os.path.abspath('')) / "testdata" / "OriginalOMI" / "omi_var.txt"
 
-@pytest.mark.filterwarnings("ignore:References for the sign of the EOFs for DOY1 have to be interpolated")
-def test_if_refdata_isfound_for_correct_spontaneous_sign_changes_in_eof_series():
-    lat = np.array([-10., 0., 10.])
-    long = np.array([0., 5.])
-    eofs = []
-    for doy in range(1, 367):
-        eof1 = np.array([1, 2, 3, 4, 5, 6]) * doy
-        eof2 = np.array([10, 20, 30, 40, 50, 60]) * doy
-        eofs.append(eof.EOFData(lat, long, eof1, eof2))
-    eofs = eof.EOFDataForAllDOYs(eofs)
-
-    try:
-        target = omi.correct_spontaneous_sign_changes_in_eof_series(eofs, True)
-    except OSError:
-        pytest.fail("Function failed with OS Error, hence the reference data has probably not been found, which points "
-                    "to an installation problem of the package: ".format(OSError))
-
 setups = [(True, 0.99, 0.99), (False, 0.999, 0.999)]
+
+
 @pytest.mark.slow  # noqa: E302
 @pytest.mark.parametrize("use_quick_temporal_filter, expectedCorr1, expectedCorr2", setups)
 @pytest.mark.skipif(not olr_data_filename.is_file(), reason="OLR data file not available")
@@ -106,23 +99,25 @@ def test_calculatePCsFromOLRWithOriginalEOFs(use_quick_temporal_filter, expected
 @pytest.mark.skipif(not eof2Dirname.is_dir(), reason="EOF2 data not available not available for comparison.")
 @pytest.mark.skipif(not origOMIPCsFilename.is_file(), reason="Original OMI PCs not available for comparison.")
 def test_completeOMIReproduction_strict_leap_year_treatment(tmp_path):
-
     errors = []
 
     # Calculate EOFs
     raw_olr = olr.load_noaa_interpolated_olr(olr_data_filename)
     shorter_olr = olr.restrict_time_coverage(raw_olr, np.datetime64('1979-01-01'), np.datetime64('2012-12-31'))
     interpolated_olr = olr.interpolate_spatial_grid_to_original(shorter_olr)
+    kiladis_pp_params = {"sign_doy1reference": True,
+                         "interpolate_eofs": True}
     eofs = omi.calc_eofs_from_olr(interpolated_olr,
-                                  sign_doy1reference=True,
-                                  interpolate_eofs=True,
-                                  strict_leap_year_treatment=True)
+                                  strict_leap_year_treatment=True,
+                                  eofs_postprocessing_type="kiladis2014",
+                                  eofs_postprocessing_params=kiladis_pp_params)
     eofs.save_all_eofs_to_npzfile(tmp_path / "test_completeOMIReproduction_strict_leap_year_treatment_EOFs.npz")
 
     # Validate EOFs against original (results are inexact but close)
     orig_eofs = eof.load_all_original_eofs_from_directory(originalOMIDataDirname)
 
-    corr_1, diff_mean_1, diff_std_1, diff_abs_percent68_1, diff_abs_percent95_1, diff_abs_percent99_1 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(orig_eofs, eofs, 1, exclude_doy366=True, percentage=False, do_print=False)
+    corr_1, diff_mean_1, diff_std_1, diff_abs_percent68_1, diff_abs_percent95_1, diff_abs_percent99_1 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(
+        orig_eofs, eofs, 1, exclude_doy366=True, percentage=False, do_print=False)
     if not np.all(corr_1 > 0.994):
         errors.append("original-validation: Correlation for EOF1 at least for one DOY too low!")
     if not np.all(diff_abs_percent99_1 < 0.0084):
@@ -130,7 +125,8 @@ def test_completeOMIReproduction_strict_leap_year_treatment(tmp_path):
     if not np.all(diff_abs_percent68_1 < 0.0018):
         errors.append("original-validation: 68% percentile for EOF1 at least for one DOY too high!")
 
-    corr_2, diff_mean_2, diff_std_2, diff_abs_percent68_2, diff_abs_percent95_2, diff_abs_percent99_2 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(orig_eofs, eofs, 2, exclude_doy366=True, percentage=False, do_print=False)
+    corr_2, diff_mean_2, diff_std_2, diff_abs_percent68_2, diff_abs_percent95_2, diff_abs_percent99_2 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(
+        orig_eofs, eofs, 2, exclude_doy366=True, percentage=False, do_print=False)
     if not np.all(corr_2 > 0.993):
         errors.append("original-validation: Correlation for EOF2 at least for one DOY too low!")
     if not np.all(diff_abs_percent99_2 < 0.0065):
@@ -139,15 +135,20 @@ def test_completeOMIReproduction_strict_leap_year_treatment(tmp_path):
         errors.append("original-validation: 68% percentile for EOF2 at least for one DOY too high!")
 
     # Validate explained variance against original (results are inexact but close)
-    orig_explained_variance_1, orig_explained_variance_2 = mjoindices.evaluation_tools.load_omi_explained_variance(original_omi_explained_variance_file)
+    orig_explained_variance_1, orig_explained_variance_2 = mjoindices.evaluation_tools.load_omi_explained_variance(
+        original_omi_explained_variance_file)
 
-    corr_var1, diff_mean_var1, diff_std_var1, diff_vec_var1, diff_abs_percent68_var1, diff_abs_percent95_var1, diff_abs_percent99_var1 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(orig_explained_variance_1, eofs.explained_variance1_for_all_doys(), do_print=False, exclude_doy366=True)
+    corr_var1, diff_mean_var1, diff_std_var1, diff_vec_var1, diff_abs_percent68_var1, diff_abs_percent95_var1, diff_abs_percent99_var1 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(
+        orig_explained_variance_1, eofs.explained_variance1_for_all_doys(), do_print=False, exclude_doy366=True)
     if not diff_std_var1 < 0.0007:
-        errors.append("original-validation: Std.Dev. of the difference of both explained variances for EOF1 is to too high!")
+        errors.append(
+            "original-validation: Std.Dev. of the difference of both explained variances for EOF1 is to too high!")
     if not diff_abs_percent99_var1 < 0.0013:
-        errors.append("original-validation: 99% percentile of the difference of both explained variances for EOF1 is to too high!")
+        errors.append(
+            "original-validation: 99% percentile of the difference of both explained variances for EOF1 is to too high!")
     if not diff_abs_percent68_var1 < 0.0007:
-        errors.append("original-validation: 68% percentile of the difference of both explained variances for EOF1 is to too high!")
+        errors.append(
+            "original-validation: 68% percentile of the difference of both explained variances for EOF1 is to too high!")
 
     corr_var2, diff_mean_var2, diff_std_var2, diff_vec_var2, diff_abs_percent68_var2, diff_abs_percent95_var2, diff_abs_percent99_var2 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(
         orig_explained_variance_2, eofs.explained_variance2_for_all_doys(), do_print=False, exclude_doy366=True)
@@ -183,7 +184,8 @@ def test_completeOMIReproduction_strict_leap_year_treatment(tmp_path):
     # reference values. So do the same with the testing target pcs.
     pcs = pc.load_pcs_from_txt_file(pc_temp_file)
 
-    tempa, tempb, corr_pc1, diff_mean_pc1, diff_std_pc1, diff_ts_abs_pc1, diff_abs_percent68_pc1, diff_abs_percent95_pc1, diff_abs_percent99_pc1 = mjoindices.evaluation_tools.calc_timeseries_agreement(orig_pcs.pc1, orig_pcs.time, pcs.pc1, pcs.time, exclude_doy366=True, do_print=False)
+    tempa, tempb, corr_pc1, diff_mean_pc1, diff_std_pc1, diff_ts_abs_pc1, diff_abs_percent68_pc1, diff_abs_percent95_pc1, diff_abs_percent99_pc1 = mjoindices.evaluation_tools.calc_timeseries_agreement(
+        orig_pcs.pc1, orig_pcs.time, pcs.pc1, pcs.time, exclude_doy366=True, do_print=False)
     if not corr_pc1 > 0.998:
         errors.append("original-validation: Correlation for PC1 timeseries is to too low!")
     if not diff_std_pc1 < 0.0449:
@@ -214,9 +216,11 @@ def test_completeOMIReproduction_strict_leap_year_treatment(tmp_path):
     if not diff_std_strength < 0.0105:
         errors.append("original-validation: Std.Dev. of the difference of both strength timeseries is to too high!")
     if not diff_abs_percent99_strength < 0.0350:
-        errors.append("original-validation: 99% percentile of the difference of both strength timeseries is to too high!")
+        errors.append(
+            "original-validation: 99% percentile of the difference of both strength timeseries is to too high!")
     if not diff_abs_percent68_strength < 0.0081:
-        errors.append("original-validation: 68% percentile of the difference of both strength timeseries is to too high!")
+        errors.append(
+            "original-validation: 68% percentile of the difference of both strength timeseries is to too high!")
 
     # Validate PCs against mjoindices own reference (results should be equal)
     mjoindices_reference_pcs = pc.load_pcs_from_txt_file(mjoindices_reference_pcs_filename_strict)
@@ -236,142 +240,143 @@ def test_completeOMIReproduction_strict_leap_year_treatment(tmp_path):
 @pytest.mark.skipif(not eof2Dirname.is_dir(), reason="EOF2 data not available not available for comparison.")
 @pytest.mark.skipif(not origOMIPCsFilename.is_file(), reason="Original OMI PCs not available for comparison.")
 def test_completeOMIReproduction(tmp_path):
+    errors = []
 
-        errors = []
+    # Calculate EOFs
+    raw_olr = olr.load_noaa_interpolated_olr(olr_data_filename)
+    shorter_olr = olr.restrict_time_coverage(raw_olr, np.datetime64('1979-01-01'), np.datetime64('2012-12-31'))
+    interpolated_olr = olr.interpolate_spatial_grid_to_original(shorter_olr)
+    kiladis_pp_params = {"sign_doy1reference": True,
+                         "interpolate_eofs": True}
+    eofs = omi.calc_eofs_from_olr(interpolated_olr,
+                                  strict_leap_year_treatment=False,
+                                  eofs_postprocessing_type="kiladis2014",
+                                  eofs_postprocessing_params=kiladis_pp_params)
+    eofs.save_all_eofs_to_npzfile(tmp_path / "test_completeOMIReproduction_EOFs.npz")
 
-        # Calculate EOFs
-        raw_olr = olr.load_noaa_interpolated_olr(olr_data_filename)
-        shorter_olr = olr.restrict_time_coverage(raw_olr, np.datetime64('1979-01-01'), np.datetime64('2012-12-31'))
-        interpolated_olr = olr.interpolate_spatial_grid_to_original(shorter_olr)
-        eofs = omi.calc_eofs_from_olr(interpolated_olr,
-                                      sign_doy1reference=True,
-                                      interpolate_eofs=True,
-                                      strict_leap_year_treatment=False)
-        eofs.save_all_eofs_to_npzfile(tmp_path / "test_completeOMIReproduction_EOFs.npz")
+    # Validate EOFs against original (results are inexact but close)
+    orig_eofs = eof.load_all_original_eofs_from_directory(originalOMIDataDirname)
 
-        # Validate EOFs against original (results are inexact but close)
-        orig_eofs = eof.load_all_original_eofs_from_directory(originalOMIDataDirname)
+    corr_1, diff_mean_1, diff_std_1, diff_abs_percent68_1, diff_abs_percent95_1, diff_abs_percent99_1 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(
+        orig_eofs, eofs, 1, exclude_doy366=False, percentage=False, do_print=False)
+    if not np.all(corr_1 > 0.994):
+        errors.append("original-validation: Correlation for EOF1 at least for one DOY too low!")
+    if not np.all(diff_abs_percent99_1 < 0.0084):
+        errors.append("original-validation: 99% percentile for EOF1 at least for one DOY too high!")
+    if not np.all(diff_abs_percent68_1 < 0.0018):
+        errors.append("original-validation: 68% percentile for EOF1 at least for one DOY too high!")
 
-        corr_1, diff_mean_1, diff_std_1, diff_abs_percent68_1, diff_abs_percent95_1, diff_abs_percent99_1 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(
-            orig_eofs, eofs, 1, exclude_doy366=False, percentage=False, do_print=False)
-        if not np.all(corr_1 > 0.994):
-            errors.append("original-validation: Correlation for EOF1 at least for one DOY too low!")
-        if not np.all(diff_abs_percent99_1 < 0.0084):
-            errors.append("original-validation: 99% percentile for EOF1 at least for one DOY too high!")
-        if not np.all(diff_abs_percent68_1 < 0.0018):
-            errors.append("original-validation: 68% percentile for EOF1 at least for one DOY too high!")
+    corr_2, diff_mean_2, diff_std_2, diff_abs_percent68_2, diff_abs_percent95_2, diff_abs_percent99_2 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(
+        orig_eofs, eofs, 2, exclude_doy366=False, percentage=False, do_print=False)
+    if not np.all(corr_2 > 0.993):
+        errors.append("original-validation: Correlation for EOF2 at least for one DOY too low!")
+    if not np.all(diff_abs_percent99_2 < 0.0065):
+        errors.append("original-validation: 99% percentile for EOF2 at least for one DOY too high!")
+    if not np.all(diff_abs_percent68_2 < 0.0018):
+        errors.append("original-validation: 68% percentile for EOF2 at least for one DOY too high!")
 
-        corr_2, diff_mean_2, diff_std_2, diff_abs_percent68_2, diff_abs_percent95_2, diff_abs_percent99_2 = mjoindices.evaluation_tools.calc_comparison_stats_for_eofs_all_doys(
-            orig_eofs, eofs, 2, exclude_doy366=False, percentage=False, do_print=False)
-        if not np.all(corr_2 > 0.993):
-            errors.append("original-validation: Correlation for EOF2 at least for one DOY too low!")
-        if not np.all(diff_abs_percent99_2 < 0.0065):
-            errors.append("original-validation: 99% percentile for EOF2 at least for one DOY too high!")
-        if not np.all(diff_abs_percent68_2 < 0.0018):
-            errors.append("original-validation: 68% percentile for EOF2 at least for one DOY too high!")
+    # Validate explained variance against original (results are inexact but close)
+    orig_explained_variance_1, orig_explained_variance_2 = mjoindices.evaluation_tools.load_omi_explained_variance(
+        original_omi_explained_variance_file)
 
-        # Validate explained variance against original (results are inexact but close)
-        orig_explained_variance_1, orig_explained_variance_2 = mjoindices.evaluation_tools.load_omi_explained_variance(
-            original_omi_explained_variance_file)
+    corr_var1, diff_mean_var1, diff_std_var1, diff_vec_var1, diff_abs_percent68_var1, diff_abs_percent95_var1, diff_abs_percent99_var1 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(
+        orig_explained_variance_1, eofs.explained_variance1_for_all_doys(), do_print=False, exclude_doy366=False)
+    if not diff_std_var1 < 0.0008:
+        errors.append(
+            "original-validation: Std.Dev. of the difference of both explained variances for EOF1 is to too high!")
+    if not diff_abs_percent99_var1 < 0.0017:
+        errors.append(
+            "original-validation: 99% percentile of the difference of both explained variances for EOF1 is to too high!")
+    if not diff_abs_percent68_var1 < 0.0009:
+        errors.append(
+            "original-validation: 68% percentile of the difference of both explained variances for EOF1 is to too high!")
 
-        corr_var1, diff_mean_var1, diff_std_var1, diff_vec_var1, diff_abs_percent68_var1, diff_abs_percent95_var1, diff_abs_percent99_var1 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(
-            orig_explained_variance_1, eofs.explained_variance1_for_all_doys(), do_print=False, exclude_doy366=False)
-        if not diff_std_var1 < 0.0008:
-            errors.append(
-                "original-validation: Std.Dev. of the difference of both explained variances for EOF1 is to too high!")
-        if not diff_abs_percent99_var1 < 0.0017:
-            errors.append(
-                "original-validation: 99% percentile of the difference of both explained variances for EOF1 is to too high!")
-        if not diff_abs_percent68_var1 < 0.0009:
-            errors.append(
-                "original-validation: 68% percentile of the difference of both explained variances for EOF1 is to too high!")
+    corr_var2, diff_mean_var2, diff_std_var2, diff_vec_var2, diff_abs_percent68_var2, diff_abs_percent95_var2, diff_abs_percent99_var2 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(
+        orig_explained_variance_2, eofs.explained_variance2_for_all_doys(), do_print=False, exclude_doy366=False)
+    if not diff_std_var2 < 0.0008:
+        errors.append(
+            "original-validation: Std.Dev. of the difference of both explained variances for EOF2 is to too high!")
+    if not diff_abs_percent99_var2 < 0.0018:
+        errors.append(
+            "original-validation: 99% percentile of the difference of both explained variances for EOF2 is to too high!")
+    if not diff_abs_percent68_var2 < 0.001:
+        errors.append(
+            "original-validation: 68% percentile of the difference of both explained variances for EOF2 is to too high!")
 
-        corr_var2, diff_mean_var2, diff_std_var2, diff_vec_var2, diff_abs_percent68_var2, diff_abs_percent95_var2, diff_abs_percent99_var2 = mjoindices.evaluation_tools.calc_comparison_stats_for_explained_variance(
-            orig_explained_variance_2, eofs.explained_variance2_for_all_doys(), do_print=False, exclude_doy366=False)
-        if not diff_std_var2 < 0.0008:
-            errors.append(
-                "original-validation: Std.Dev. of the difference of both explained variances for EOF2 is to too high!")
-        if not diff_abs_percent99_var2 < 0.0018:
-            errors.append(
-                "original-validation: 99% percentile of the difference of both explained variances for EOF2 is to too high!")
-        if not diff_abs_percent68_var2 < 0.001:
-            errors.append(
-                "original-validation: 68% percentile of the difference of both explained variances for EOF2 is to too high!")
+    # Validate EOFs against mjoindices own reference (results should be equal)
+    mjoindices_reference_eofs = eof.restore_all_eofs_from_npzfile(mjoindices_reference_eofs_filename)
+    for idx, target_eof in enumerate(eofs.eof_list):
+        if not mjoindices_reference_eofs.eof_list[idx].close(target_eof):
+            errors.append("mjoindices-reference-validation: EOF data at index %i is incorrect" % idx)
 
-        # Validate EOFs against mjoindices own reference (results should be equal)
-        mjoindices_reference_eofs = eof.restore_all_eofs_from_npzfile(mjoindices_reference_eofs_filename)
-        for idx, target_eof in enumerate(eofs.eof_list):
-            if not mjoindices_reference_eofs.eof_list[idx].close(target_eof):
-                errors.append("mjoindices-reference-validation: EOF data at index %i is incorrect" % idx)
+    # Calculate PCs
+    raw_olr = olr.load_noaa_interpolated_olr(olr_data_filename)
+    pcs = omi.calculate_pcs_from_olr(raw_olr,
+                                     eofs,
+                                     np.datetime64("1979-01-01"),
+                                     np.datetime64("2018-08-28"),
+                                     use_quick_temporal_filter=False)
+    pc_temp_file = tmp_path / "test_completeOMIReproduction_PCs.txt"
+    pcs.save_pcs_to_txt_file(pc_temp_file)
 
-        # Calculate PCs
-        raw_olr = olr.load_noaa_interpolated_olr(olr_data_filename)
-        pcs = omi.calculate_pcs_from_olr(raw_olr,
-                                         eofs,
-                                         np.datetime64("1979-01-01"),
-                                         np.datetime64("2018-08-28"),
-                                         use_quick_temporal_filter=False)
-        pc_temp_file = tmp_path / "test_completeOMIReproduction_PCs.txt"
-        pcs.save_pcs_to_txt_file(pc_temp_file)
+    # Validate PCs  against original (results are inexact but close)
+    orig_pcs = pc.load_original_pcs_from_txt_file(origOMIPCsFilename)
+    # Reload pcs instead of using the calculated ones, because the saving routine has truncated some decimals of the
+    # reference values. So do the same with the testing target pcs.
+    pcs = pc.load_pcs_from_txt_file(pc_temp_file)
 
-        # Validate PCs  against original (results are inexact but close)
-        orig_pcs = pc.load_original_pcs_from_txt_file(origOMIPCsFilename)
-        # Reload pcs instead of using the calculated ones, because the saving routine has truncated some decimals of the
-        # reference values. So do the same with the testing target pcs.
-        pcs = pc.load_pcs_from_txt_file(pc_temp_file)
+    tempa, tempb, corr_pc1, diff_mean_pc1, diff_std_pc1, diff_ts_abs_pc1, diff_abs_percent68_pc1, diff_abs_percent95_pc1, diff_abs_percent99_pc1 = mjoindices.evaluation_tools.calc_timeseries_agreement(
+        orig_pcs.pc1, orig_pcs.time, pcs.pc1, pcs.time, exclude_doy366=False, do_print=False)
+    if not corr_pc1 > 0.998:
+        errors.append("original-validation: Correlation for PC1 timeseries is to too low!")
+    if not diff_std_pc1 < 0.0458:
+        errors.append("original-validation: Std.Dev. of the difference of both PC1 timeseries is to too high!")
+    if not diff_abs_percent99_pc1 < 0.157:
+        errors.append(
+            "original-validation: 99% percentile of the difference of both PC1 timeseries is to too high!")
+    if not diff_abs_percent68_pc1 < 0.0327:
+        errors.append(
+            "original-validation: 68% percentile of the difference of both PC1 timeseries is to too high!")
 
-        tempa, tempb, corr_pc1, diff_mean_pc1, diff_std_pc1, diff_ts_abs_pc1, diff_abs_percent68_pc1, diff_abs_percent95_pc1, diff_abs_percent99_pc1 = mjoindices.evaluation_tools.calc_timeseries_agreement(
-            orig_pcs.pc1, orig_pcs.time, pcs.pc1, pcs.time, exclude_doy366=False, do_print=False)
-        if not corr_pc1 > 0.998:
-            errors.append("original-validation: Correlation for PC1 timeseries is to too low!")
-        if not diff_std_pc1 < 0.0458:
-            errors.append("original-validation: Std.Dev. of the difference of both PC1 timeseries is to too high!")
-        if not diff_abs_percent99_pc1 < 0.157:
-            errors.append(
-                "original-validation: 99% percentile of the difference of both PC1 timeseries is to too high!")
-        if not diff_abs_percent68_pc1 < 0.0327:
-            errors.append(
-                "original-validation: 68% percentile of the difference of both PC1 timeseries is to too high!")
+    tempa, tempb, corr_pc2, diff_mean_pc2, diff_std_pc2, diff_ts_abs_pc2, diff_abs_percent68_pc2, diff_abs_percent95_pc2, diff_abs_percent99_pc2 = mjoindices.evaluation_tools.calc_timeseries_agreement(
+        orig_pcs.pc2, orig_pcs.time, pcs.pc2, pcs.time, exclude_doy366=False, do_print=False)
+    if not corr_pc2 > 0.998:
+        errors.append("original-validation: Correlation for PC2 timeseries is to too low!")
+    if not diff_std_pc2 < 0.0488:
+        errors.append("original-validation: Std.Dev. of the difference of both PC2 timeseries is to too high!")
+    if not diff_abs_percent99_pc2 < 0.1704:
+        errors.append(
+            "original-validation: 99% percentile of the difference of both PC2 timeseries is to too high!")
+    if not diff_abs_percent68_pc2 < 0.0353:
+        errors.append(
+            "original-validation: 68% percentile of the difference of both PC2 timeseries is to too high!")
 
-        tempa, tempb, corr_pc2, diff_mean_pc2, diff_std_pc2, diff_ts_abs_pc2, diff_abs_percent68_pc2, diff_abs_percent95_pc2, diff_abs_percent99_pc2 = mjoindices.evaluation_tools.calc_timeseries_agreement(
-            orig_pcs.pc2, orig_pcs.time, pcs.pc2, pcs.time, exclude_doy366=False, do_print=False)
-        if not corr_pc2 > 0.998:
-            errors.append("original-validation: Correlation for PC2 timeseries is to too low!")
-        if not diff_std_pc2 < 0.0488:
-            errors.append("original-validation: Std.Dev. of the difference of both PC2 timeseries is to too high!")
-        if not diff_abs_percent99_pc2 < 0.1704:
-            errors.append(
-                "original-validation: 99% percentile of the difference of both PC2 timeseries is to too high!")
-        if not diff_abs_percent68_pc2 < 0.0353:
-            errors.append(
-                "original-validation: 68% percentile of the difference of both PC2 timeseries is to too high!")
+    strength = np.sqrt(np.square(pcs.pc1) + np.square(pcs.pc2))
+    orig_strength = np.sqrt(np.square(orig_pcs.pc1) + np.square(orig_pcs.pc2))
 
-        strength = np.sqrt(np.square(pcs.pc1) + np.square(pcs.pc2))
-        orig_strength = np.sqrt(np.square(orig_pcs.pc1) + np.square(orig_pcs.pc2))
+    tempa, tempb, corr_strength, diff_mean_strength, diff_std_strength, diff_ts_abs_strength, diff_abs_percent68_strength, diff_abs_percent95_strength, diff_abs_percent99_strength = mjoindices.evaluation_tools.calc_timeseries_agreement(
+        orig_strength, orig_pcs.time, strength, pcs.time, exclude_doy366=False, do_print=False)
+    if not corr_strength > 0.9998:
+        errors.append("original-validation: Correlation for strength timeseries is to too low!")
+    if not diff_std_strength < 0.0103:
+        errors.append("original-validation: Std.Dev. of the difference of both strength timeseries is to too high!")
+    if not diff_abs_percent99_strength < 0.0341:
+        errors.append(
+            "original-validation: 99% percentile of the difference of both strength timeseries is to too high!")
+    if not diff_abs_percent68_strength < 0.0079:
+        errors.append(
+            "original-validation: 68% percentile of the difference of both strength timeseries is to too high!")
 
-        tempa, tempb, corr_strength, diff_mean_strength, diff_std_strength, diff_ts_abs_strength, diff_abs_percent68_strength, diff_abs_percent95_strength, diff_abs_percent99_strength = mjoindices.evaluation_tools.calc_timeseries_agreement(
-            orig_strength, orig_pcs.time, strength, pcs.time, exclude_doy366=False, do_print=False)
-        if not corr_strength > 0.9998:
-            errors.append("original-validation: Correlation for strength timeseries is to too low!")
-        if not diff_std_strength < 0.0103:
-            errors.append("original-validation: Std.Dev. of the difference of both strength timeseries is to too high!")
-        if not diff_abs_percent99_strength < 0.0341:
-            errors.append(
-                "original-validation: 99% percentile of the difference of both strength timeseries is to too high!")
-        if not diff_abs_percent68_strength < 0.0079:
-            errors.append(
-                "original-validation: 68% percentile of the difference of both strength timeseries is to too high!")
+    # Validate PCs against mjoindices own reference (results should be equal)
+    mjoindices_reference_pcs = pc.load_pcs_from_txt_file(mjoindices_reference_pcs_filename)
+    if not np.all(mjoindices_reference_pcs.time == pcs.time):
+        errors.append("mjoindices-reference-validation: Dates of PCs do not match.")
+    if not np.allclose(mjoindices_reference_pcs.pc1, pcs.pc1):
+        errors.append("mjoindices-reference-validation: PC1 values do not match.")
+    if not np.allclose(mjoindices_reference_pcs.pc2, pcs.pc2):
+        errors.append("mjoindices-reference-validation: PC2 values do not match.")
 
-        # Validate PCs against mjoindices own reference (results should be equal)
-        mjoindices_reference_pcs = pc.load_pcs_from_txt_file(mjoindices_reference_pcs_filename)
-        if not np.all(mjoindices_reference_pcs.time == pcs.time):
-            errors.append("mjoindices-reference-validation: Dates of PCs do not match.")
-        if not np.allclose(mjoindices_reference_pcs.pc1, pcs.pc1):
-            errors.append("mjoindices-reference-validation: PC1 values do not match.")
-        if not np.allclose(mjoindices_reference_pcs.pc2, pcs.pc2):
-            errors.append("mjoindices-reference-validation: PC2 values do not match.")
-
-        assert not errors, "errors occurred:\n{}".format("\n".join(errors))
+    assert not errors, "errors occurred:\n{}".format("\n".join(errors))
 
 
 @pytest.mark.slow
@@ -380,7 +385,6 @@ def test_completeOMIReproduction(tmp_path):
 @pytest.mark.skipif(not eof2Dirname.is_dir(), reason="EOF2 data not available not available for comparison.")
 @pytest.mark.skipif(not origOMIPCsFilename.is_file(), reason="Original OMI PCs not available for comparison.")
 def test_completeOMIReproduction_coarsegrid(tmp_path):
-
     errors = []
 
     # Calculate EOFs
@@ -389,10 +393,12 @@ def test_completeOMIReproduction_coarsegrid(tmp_path):
     coarse_lat = np.arange(-20., 20.1, 8.0)
     coarse_long = np.arange(0., 359.9, 20.0)
     interpolated_olr = olr.interpolate_spatial_grid(shorter_olr, coarse_lat, coarse_long)
+    kiladis_pp_params = {"sign_doy1reference": True,
+                         "interpolate_eofs": True}
     eofs = omi.calc_eofs_from_olr(interpolated_olr,
-                                  sign_doy1reference=True,
-                                  interpolate_eofs=True,
-                                  strict_leap_year_treatment=True)
+                                  strict_leap_year_treatment=True,
+                                  eofs_postprocessing_type="kiladis2014",
+                                  eofs_postprocessing_params=kiladis_pp_params)
     eofs.save_all_eofs_to_npzfile(tmp_path / "test_completeOMIReproduction_coarsegrid_EOFs.npz")
 
     # Validate EOFs against mjoindices own reference (results should be equal)
@@ -427,22 +433,25 @@ def test_completeOMIReproduction_coarsegrid(tmp_path):
 
 
 eofs_spec = importlib.util.find_spec("eofs")
+
+
 @pytest.mark.slow
 @pytest.mark.skipif(not olr_data_filename.is_file(), reason="OLR data file not available.")
 @pytest.mark.skipif(eofs_spec is None, reason="Optional eofs package is not available.")
 def test_completeOMIReproduction_eofs_package_strict_leap_year(tmp_path):
-
     errors = []
 
     # Calculate EOFs
     raw_olr = olr.load_noaa_interpolated_olr(olr_data_filename)
     shorter_olr = olr.restrict_time_coverage(raw_olr, np.datetime64('1979-01-01'), np.datetime64('2012-12-31'))
     interpolated_olr = olr.interpolate_spatial_grid_to_original(shorter_olr)
+    kiladis_pp_params = {"sign_doy1reference": True,
+                         "interpolate_eofs": True}
     eofs = omi.calc_eofs_from_olr(interpolated_olr,
-                                  sign_doy1reference=True,
-                                  interpolate_eofs=True,
+                                  implementation="eofs_package",
                                   strict_leap_year_treatment=True,
-                                  implementation="eofs_package")
+                                  eofs_postprocessing_type="kiladis2014",
+                                  eofs_postprocessing_params=kiladis_pp_params)
     eofs.save_all_eofs_to_npzfile(tmp_path / "test_completeOMIReproduction_strict_leap_year_treatment_EOFs.npz")
 
     # Validate EOFs against mjoindices own reference (results should be equal)
@@ -488,5 +497,54 @@ def test_preprocess_olr_warning():
                            (90., 100., 110., 120.))])
     olrmatrix *= -1.
     testdata = olr.OLRData(olrmatrix, time, lat, long)
-    with pytest.warns(UserWarning, match="OLR data apparently given in negative numbers. Here it is assumed that OLR is positive."):
+    with pytest.warns(UserWarning,
+                      match="OLR data apparently given in negative numbers. Here it is assumed that OLR is positive."):
         target = omi.preprocess_olr(testdata)
+
+
+def test_initiate_eof_post_processing():
+    errors = []
+
+    raw_eofs = eof.restore_all_eofs_from_npzfile(mjoindices_reference_eofs_filename_raw)
+
+    # Check that user receives Error if there is a typo in the pp-type
+    with pytest.raises(ValueError):
+        eofs = omi.initiate_eof_post_processing(raw_eofs, eofs_postprocessing_type="XXX",
+                                                eofs_postprocessing_params=None)
+
+    # Check no pp wanted
+    eofs = omi.initiate_eof_post_processing(raw_eofs, eofs_postprocessing_type=None,
+                                            eofs_postprocessing_params=None)
+
+    # Validate EOFs against mjoindices own reference of raw EOFs
+    for idx, target_eof in enumerate(eofs.eof_list):
+        if not raw_eofs.eof_list[idx].close(target_eof):
+            errors.append("No postprocessing wanted: EOF data at index %i is incorrect" % idx)
+
+    # Check Kiladis-PP
+    kiladis_params = {"sign_doy1reference": True,
+                      "interpolate_eofs": True,
+                      "interpolation_start_doy": 293,
+                      "interpolation_end_doy": 316}
+
+    eofs = omi.initiate_eof_post_processing(raw_eofs, eofs_postprocessing_type="kiladis2014",
+                                            eofs_postprocessing_params=kiladis_params)
+
+    # Validate EOFs against mjoindices own reference (results should be equal)
+    mjoindices_reference_eofs = eof.restore_all_eofs_from_npzfile(mjoindices_reference_eofs_filename)
+    for idx, target_eof in enumerate(eofs.eof_list):
+        if not mjoindices_reference_eofs.eof_list[idx].close(target_eof):
+            errors.append("Original Kiladis PP Test: EOF data at index %i is incorrect" % idx)
+
+    # With the following line, we simply check that the kiladis pp-routine is executable with default params.
+    #This is quick and dirty, since we do not check here if the parameters are actually applied!
+    #ToDo: Check the actual application of the default parameters
+    eofs = omi.initiate_eof_post_processing(raw_eofs, eofs_postprocessing_type="kiladis2014",
+                                            eofs_postprocessing_params=None)
+
+    # Check EOFRotation-PP
+    # ToDo: (Sarah): Add a similar test for your pp here to make sure, that it is actually executed by the initiation routine...
+    # ... I am aware of the fact that this tests the numbers as the specific unit tests in the sepearate file for your pp routine...
+    # ... However, I thin it it important to test the the right pp routine is actually executed by the initiation routine.
+
+    assert not errors, "errors occurred:\n{}".format("\n".join(errors))
